+++ voluncheer/jobboard/apps.py
@@ -1,7 +1,7 @@
 from django.apps import AppConfig
 
 
-class JobBoardConfig(AppConfig):
+class JobboardConfig(AppConfig):
     """Django app configuration for volunteer opportunities."""
     default_auto_field = "django.db.models.BigAutoField"
     name = "jobboard"
