+++ voluncheer/jobboard/models.py
@@ -2,7 +2,6 @@ from django.db import models
 
 
 class Job(models.Model):
-    id=models.IntegerField(primary_key=True)
     organization = models.ForeignKey("profiles.Organization", on_delete=models.CASCADE)
     title = models.CharField(max_length=256)
     pubdate = models.DateTimeField('date published')
