+++ voluncheer/profiles/tests.py
@@ -5,6 +5,12 @@ from profiles.models import User
 from profiles.models import UserType
 from profiles.models import Volunteer
 
+from django.test import Client
+from django.test import TestCase
+from django.urls import reverse
+
+from profiles import urls
+from profiles import views
 
 class UserTest(TestCase):
     """Test cases for the basic user type."""
@@ -67,3 +73,14 @@ class UserTest(TestCase):
         """Tests basic volunteer details."""
         self.assertEqual(self.luke.name, "Luke Skywalker")
         self.assertEqual(self.vader.name, "Darth Vader")
+        
+
+# Create your tests here.
+class ProfilesTestCase(TestCase):
+    def test_page_loads(self):
+        c = Client()
+        for url in urls.urlpatterns:
+            appNameAndUrl = urls.app_name+":"+url.name
+            print(appNameAndUrl)
+            response = c.get(reverse(appNameAndUrl))
+            self.assertEqual(response.status_code, 200)
