+++ voluncheer/requirements.txt
@@ -4,4 +4,3 @@ django_debug_toolbar==3.8.1
 Django==3.2.8
 pytz==2022.7.1
 sqlparse==0.4.3
-pysqlite3-binary=0.5.0
