+++ voluncheer/voluncheer/urls.py
@@ -32,6 +32,6 @@ urlpatterns = [
 
     # Unimplemented urls.
     # path("chat/", include("chatroom.urls")),
-     path("jobboard/", include("jobboard.urls")),
+    path("jobboard/", include("jobboard.urls")),
     path("map/", include("map.urls")),
 ]
